feat: add description for products in catalog page
