+++ src/api/BuildClient.ts
@@ -41,5 +41,5 @@ export const ctpClient = new ClientBuilder()
   .withProjectKey(projectKey) // .withProjectKey() is not required if the projectKey is included in authMiddlewareOptions
   .withClientCredentialsFlow(authMiddlewareOptions)
   .withHttpMiddleware(httpMiddlewareOptions)
-  // .withLoggerMiddleware() // Include middleware for logging
+  // .withLoggerMiddleware() // @note используем чтобы выводить в консоль запросы отправляемые commercetools/sdk
   .build();
