+++ src/components/ProductsCategory.module.scss
@@ -1,16 +1,22 @@
 @import '../styles/global';
 
+.main {
+  display: flex;
+  justify-content: space-between;
+  gap: 10px;
+}
+
 .category {
   display: flex;
   flex-wrap: wrap;
-  justify-content: space-between;
+  justify-content: space-around;
   width: 90vw;
   margin: 20px 0px;
-  padding-bottom: 0px;
   > * {
     height: 140px;
     width: 140px;
     padding-top: 5px;
+    margin: 10px 0;
     text-align: center;
     align-items: flex-end;
     background-size: cover;
