+++ src/styles/_constants.scss
@@ -8,4 +8,4 @@ $color-green: #168d16;
 
 $tablet-width: 780px;
 $mobile-width: 500px;
-$hamburger-width: 640px;
+$hamburger-width: 730px;
